# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Punjabi Palace API",
        "description": "This is the API documentation for the Restaurant management application.",
        "contact": {
            "email": "ravikumar@adzguruexports.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stocks": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get all stocks",
                "operationId": "1d81e97f8764d93a370254e42ff7542d",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved stocks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "stock_item_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "quantity": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "uom_id": {
                                                "type": "integer"
                                            },
                                            "unit": {
                                                "type": "string"
                                            },
                                            "minimum_qty": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "unit_price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create a new stock",
                "operationId": "2809028990ca5a2f8f971c7e9c086475",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stock_item_id",
                                    "name",
                                    "quantity",
                                    "uom_id",
                                    "unit",
                                    "minimum_qty"
                                ],
                                "properties": {
                                    "stock_item_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "uom_id": {
                                        "type": "integer"
                                    },
                                    "unit": {
                                        "type": "string"
                                    },
                                    "minimum_qty": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock created successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/stocks/{id}": {
            "put": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Update a stock",
                "operationId": "814647a8a786738fdf7be685c9059e32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stock_item_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "uom_id": {
                                        "type": "integer"
                                    },
                                    "unit": {
                                        "type": "string"
                                    },
                                    "minimum_qty": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "unit_price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock updated successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete a stock",
                "operationId": "5aef6f6787f39978fe839320abe82d97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock deleted successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/stock-items": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get all stock items",
                "operationId": "dae1d6e4c54acc9d8eac5855452124af",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved stock items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "item_code": {
                                                "type": "string"
                                            },
                                            "item_name": {
                                                "type": "string"
                                            },
                                            "item_category_id": {
                                                "type": "integer"
                                            },
                                            "uom_id": {
                                                "type": "integer"
                                            },
                                            "uom_name": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create a new stock item",
                "operationId": "19daf9d0d3e57a3de8594edde439dfad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_code",
                                    "item_name",
                                    "item_category_id",
                                    "uom_id"
                                ],
                                "properties": {
                                    "item_code": {
                                        "type": "string"
                                    },
                                    "item_name": {
                                        "type": "string"
                                    },
                                    "item_category_id": {
                                        "type": "integer"
                                    },
                                    "uom_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}